difftest: add a special field to instr commit
This commit adds a special field to instruction commit packet. For now,
it is used to indicate that the instruction is a fused instruction. In
the future, we will move isRVC, scFailed, moveElim, etc into this field.

When the instruction is fused, we need to let REF commit one more
instruction.