Add preliminary support for batch mode
For now, we only implement a 2D array to store the difftest trace.
The sync logic should be further implemented to support the real
batch mode.

The sync may reuse other styles such as DPI-C. However, it should
be carefully implemented to act like a real batch data transfer.
We don't know how it looks like now so we are leaving it empty.

The C++ difftest files also need to support batch mode with an array
of DiffTrace as the input. Actually we already have the DiffTrace
abstraction, the only thing to do is to avoid calling simulation ticks
when dealing with the batch data and resume simulation after it.