DPIC: assign enable with valid
Previously we are using the valid bit in the generated C++ function
to decide whether to update other fields. However, this brings both
performance and functional issues.

When valid is set to false in the hardware, we still need to call
the DPI-C function, causing performance overhead. Besides, we still
update the valid field in C++. However, the index may be invalid at
this time and brings out-of-range memory accesses.

Now we move the valid condition to the enable bit in DPI-C module
wrapper. The simulator should avoid calling the DPI-C function when
not necessary.