Gateway,top: remove batch Buffer and select, modify step logic.
Blocking time of Palladium Hardware depends on number of DPICs. There
is no need to store DPICs and pass them at the same time because that will
not changes its number. Instead, it consumes far more gates.

In simulation platform such as VCS and Palladium, the order of DPICs
at same cycle is not fixed, so we delay step by one clock to ensure
transmission is done. However, it means buffer read (when step) may be
conflicted with write. So we add an extra pos to dut_buffer, ensuring
no pos will be read and write at the same time.

We move the implementation of step delay from top to gateway.