Use DPIC-task to return deferred result
Before this commit, we fetch simv_result at 5000 cycles. Only
when the result is not zero, it works and stop simulation.

Now we use DPIC task to implement the return of simv_result when not
zero, so only one DPIC task will be used to stop simulation.
That will reduce the cost of fetch and get result return faster.