Fix Palladium
We add Macros to turn off DPIC on MemRWHelper and SimJTAG without
SYNTHESIS. That means we can reduce number of synchronizations if
needed. The new Macros is defined in palladium.mk by default.

To add build-in decleration of system task for Palladium, we need
to add $.

We declare fwrite in tb_top as Gfifo to reduce synchronizations and
speed up simulation.