Use DPIC-function to return deferred result (#250)
Before this commit, we fetch simv_result at 5000 cycles. Only
when the result is not zero, it works and stop simulation.

Now we use DPIC function to implement the return of simv_result when not
zero, so only one DPIC function will be used to stop simulation.
That will reduce the cost of fetch and get result return faster.