Allow LogPerfControl to be reused if necessary
Since https://github.com/chipsalliance/chisel/pull/3753, the users
are able to check whether a Data is visible in the current context.
To avoid a large number of duplicated LogPerfControl modules, we
allow it to be reused when some previous instance is still visible
to the current context.

Call `LogPerfControl.reuse(DataMirror.isVisible)` instead of pure
`apply()` to allow the reuse.

However, this is still a temporary workaround. In the future, we
need some clean methods to extract the simulation environment and
access these signals during simulation, possibly via XMR.