Batch, DPIC: reshape code structure
For Batch, we add abstract class DPICBase for common methods and
template. DPIC/DPICBatch only need to override little logic to
reuse it.
For DPIC, we seperate io and control, because both Single/Batch will
use same control interface, but different io interface. We don't
need to declare extra bundle for that.