InternalStep: pack Data transmit and Step control
* Configurable batchByteLen:
For better performace without GFIFO. In palladium, we can pass
maximum 4000 byte in Gfifo Func (enable isNonBlock), and 8000
byte in Other Func.
With configurable byteLen, speed will increase from 70 K to 98 K
without GFIFO.

* Interval step: pack Data transmit and Step control
Previous when batch is not supported, we use multiple DPIC func to
transmit DiffState. Since the DPIC order in one beat is uncertain,
we use the step to mark the transmission is finished.

With batch mode, data will be transmit at single func, so we can
also pack step to reduce sync times. It can reduce DPIC calls to
half. Speed will increase from 95 K to 98 K without GFIFO, 26 K
to 27.5 K with GFIFO.

When IntervalStep is enable, step in TOP IO will be default 0. We
hold this unused signal for fixed TOP IO interface.

By the way, when the feature is supported. DUT will only use a
single port to interface with REF. This helps us to migrate Difftest for other platform.