Add the reflective DataMirror for future methods
This commit brings a DataMirror in the difftest.util package. It
now includes a method called isVisible which returns the visibility
of a Data object, which was originally brought into Chisel 6.1.0 for
the first time.

Since this method does not exist in older versions of Chisel, and
DiffTest must be compatible with legacy versions of Chisel, we use
Scala reflection features to load this method dynamically. If not
found, this method always returns false.