vcs: support track pc
when the commit PC is same value as the tracking PC, print the clock cycle.
Sometimes the result is inaccurate for branch instruction and fused instruction.
When multiple instructions are committed in a single cycle, the commit PC is
inferred from the dut->commit[0].pc at the beginning of each cycle and the number
of instructions has been committed in the current cycle.