emu: update the stack size only for legacy Verilator
Verilator v5.026 comes with an automated stack resizer for various
verilated models. The stack size will be set implicitly in the
constructor of the verilated design based on an estimated stack size
requirement.

Thus, we comment out the `setrlimit` lines for newer version of
Verilator. They will use the estimated stack size instead of the
predetermined constant value.