Makefile: set difftest_verilog the default target (#368)
This commit changes the default target of make to difftest_verilog.

The previous default target is emu. However, emu is for only the Verilator
target, one of the three supported simulation platforms.

To make the default target clearer, we change the default target to
difftest_verilog. It generates example difftest design-under-test
interfaces and C++ files. Using this main target, user may connect
their own design (maybe in Verilog) to difftest for simulation.