Avoid warning messages when ccache does not exist (#369)
verilator.mk checks the existence of ccache and use it when possible.
However, on some OS platforms, such as CentOS, this checker may cause
warning messages if ccache does not exist. Since this checker is always
called even when the user is not building Verilator targets, the warning
message seems confused for users.

We avoid this message by redirecting the stderr to /dev/null.