spike-dasm: performance improvement
We remove unnecessary extra logic to dasm the instruction string.
The execution speed is improved by 1.78X.

We also add redirection for stderr when checking the existence of
spike-dasm. This should avoid the not-found warning when running
the simulation.

After this change, the user may not be aware of the spike-dasm plugin
for disassembly the instruction. Therefore, we update the README as
well.