difftest: allow early store events for the current instruction (#379)
Previously we require store events to be updated after the store
instruction is committed. This commit further allows early store
events for the current to-be-committed instruction(s). The next
commit event, which contains a group of instructions will flush
the store queue and check all updated stores. It's worth noting
store events are allowed only for the current pending commit
instructions. Store events for future instructions (not committed
in the next commit event) will make the co-simulation fail.

Co-authored-by: Anzooooo <anzoso@outlook.com>