verilator.mk: Add support for PGO
This commit add PGO_CFLAGS and PGO_LDFLAGS when calling verilator to
generate Makefile. And add script to build PGO for both GCC and LLVM.

The following is the usage of PGO:

```shell
make emu MFC=1 EMU_THREADS=16 NUM_CORES=1 PGO_WORKLOAD=`realpath ready-to-run/coremark-2-iteration.bin` LLVM_PROFDATA=llvm-profdata -j `nproc`
```

Note: The LLVM_PROFDATA is hard to detect if multiple versions of LLVM
are installed. An example is that on Debian, we can install different
versions of LLVM and use a version with suffixes like `clang++-18`. The
LLVM_PROFDATA should be set to the correct version like
`llvm-profdata-18` in this case.

If the compiler of verilator is GCC, we should not use LLVM_PROFDATA:

```shell
make emu MFC=1 EMU_THREADS=16 NUM_CORES=1 PGO_WORKLOAD=`realpath ready-to-run/coremark-2-iteration.bin` -j `nproc`
```

Signed-off-by: Yangyu Chen <cyy@cyyself.name>