verilator.mk: Add OPT_FAST option and refractor build flow
This commit adds an OPT_FAST option for choosing the optimization
level for verilator-generated C sources, which is useful to address
the slow compile time issue on clang-18. And also addressed #395.

The changes to address #395 are:

1. Remove the dependence of `$(EMU_MK)` on build_emu_local

A required chain exists in the current ventilator.mk:

`build_emu_local` -> `$(EMU_MK)` -> `$(EMU_DEPS)` -> `$(EMU_CXXFILES)` -> `$(SIM_CXXFILES)`

However, when we have `WITH_CHISELDB=1` which is default, we will do
`SIM_CXXFILES += $(BUILD_DIR)/chisel_db.cpp` then, `chisel_db.cpp`
will be required by build_emu_local. If this file does not exist,
the make command will fail. However, we pass `WITH_CHISELDB` from
another Makefile, if this does not pass to ssh, building on another
host will fail. If some arguments like this all passed to ssh, the
ssh command will be very long. And there is no need for other hosts
to compile just to resolve the correct requirements of the
`$(EMU_MK)`.

Thus, I remove the dependence of `$(EMU_MK)` on build_emu_local.
It's safe since the build_emu_local will only be called on the
`$(EMU)` target, and the dependence has already been resolved.

2. Use localhost for PGO workload running

PGO requires some parameters set in the Makefile, and the workload
running time on remote server with huge number of cores does not
reflect the real performance of the program using just a few of
`EMU_THREADS`. Also, if we need to run PGO on remote server, we
have to pass all the PGO parameters such as `PGO_WORKLOAD` to
remote server, which makes the code hard to maintain.

Thus, I use localhost for PGO workload running now.

3. Some refractors on the verilator.mk

To reduce the maintenance cost, we need to explicitly define what
veriable should be passed when making the `$(EMU)` target. Thus,
I rename `build_emu_local` with `build_emu` and make local build
command and ssh command closer just in the if block. So one day if
a developer needs some new arguments, he can easily find the place
to add them.

Signed-off-by: Yangyu Chen <cyy@cyyself.name>