verilator.mk: Sync BUILD_DIR and EMU_DIR for PGO
When building PGO Verilator EMU on a remote server with shared NFS
mounted with async. After compiling on a remote server, the dirty
medication on filesystem may not been flushed to the NFS server,
or (and) the localhost has not received a cache coherence snoop
request from the NFS server. In this case, executing emu on localhost
may fail because the file cannot be found. This commit addressed
this by adding sync to the build_dir and emu_dir. So we have the
right directory metadata synced between the build server and
localhost. I tried to build a PGO verilator emu several times on a
remote server. It no longer fails.

Signed-off-by: Yangyu Chen <cyy@cyyself.name>