difftest: support print CPI to specific file (#400)
When specify ”OUTPUT_CPI_TO_FILE“ with a file name, difftest will print [coreid,CPI] to a temporary file at the end of the simulation. This information is mainly used to support the generation of multi-core checkpoints, which needs RTL CPI info to guide qemu simulation. Now this feature only supported in VCS flow.