util: add support for synthesizable assertions
We add a VerificationExtractor transform with an annotation named
VerificationExtractorSink to support synthesizable assertions.

First, the user should annotate a sink Bool variable to let the
transform know that it should do wiring for all assertions. Second,
the transform collects all assertions (firrtl.ir.Verification) from
the circuit and adds SourceAnnotation for wiring. Third, in the sink
module, the transform creates Cat and orR nodes to compute a Bool
value for the annotated assertion sink.

The user should 1) annotate the sink Bool using the function
VerificationExtractor.sink, and 2) add VerificationExtractor to the
build dependecy as one of the transforms.

This transform does not support Chisel 3.x now and in the future.
To allow compiling the code using Chisel 3.x, we use Scala reflection
to create DefRegisterWithReset nodes.

It's worth noting that, this transform only creates a 1-bit value
for detecting the assertions. It does not support any debugging
methodologies to expose the 1-bit value to the simulation environment.
The debugging method should depend on the simulation platform.