verification: remove some clock/reset dependencies
This commit partially removes the dependencies on clock and reset
ports for the verification extractor. The sink will no longer holds
the assertion conditions once it is asserted. The user should use
its own clock to sample it and hold it when necessary.

Previous dependencies require both clock and reset ports for the
orR sink, and reset ports for assert source modules.

This commit adds the clock and reset for sink annotation and thus
removes the dependencies. It also removes the reset requirement
for assert source modules.