Validate: recover needUpdate for TrapEvent
TrapEvent may not raise hasTrap when any other bundle valid, so we
recover needUpdate to mark there is valid TrapEvent at this cycle.

In validate, we use getValid to mark valid for single bundle, and
needUpdate to mark valid for this cycle. We will transmit trapEvent
when hasTrap, or other bundles valid.