feat: add design version info at runtime
Recently, we have added version info of the design code (the commit id) in the XiangShan RTL code. This patch adds the version info in the emu/simv program. Now the emu/simv program will print the commit id of design when initializing.

This patch also resolves the issue where the emu build time would not update when re-compiling. Now, with each build, `version.h` is re-generated, causing `common.cpp`, the source code file which contains printing logic, to be recompiled, which subsequently triggers a relink of `emu`. This process takes a few seconds, which is longer than before, but still acceptable.