feat(difftest): support parsing recursive bundles
We support pattern matching of Bundle for elements of DifftestBundle.
This allows the usage of Bundle when defining a DifftestBundle.

Also, the previous DPI-C implementation ignores bundles and assumes all
top-level signals are either Vec or UInt. This commit fixes it by
reusing the recursive function defined in DifftestBundle to get all
elements in UInt. This ensures the same functionality between DPI-C and
DifftestBundle.