fix(Makefile): set the default REF to Nemu (#528)
This commits add the default REF variable with value of Nemu.

We intended to record the REF_HOME at compile-time with the REF_HOME
macro passed as a C++ flag. However, this detection is wrapped in
the `ifneq ($(REF),)` checker and fails to capture NEMU_HOME if the
user does not explicitly specify `REF=Nemu`.

We also remove the default REF_PROXY variable in config.h.