FPGA: support simulate fpga-host with dut (#624)
This change supports software simulation of FPGA-host interacting
with DUT in a xdma-like way. We use a simv process to simulate DUT
and send difftest_data with xdma-like pcie. The simv process will
interacting with fpga-host by share memory acting as xdma device.

Code of the xdma_ctrl.v is modified from the code of fengkehan
<fengkehan@bosc.ac.cn> and fix some bugs.