fpga: optimize timing and area with BRAM (#655)
Previously, we used registers to implement the multi-buffer sending module, which resulted in poorer resource utilization and timing performance. 

This change adopts BRAM for optimizing the timing and area. The LUT occupancy has been reduced from 17W to 2W, while the BRAM occupancy has increased to 448 (11% of the total resources of VU19P). And the timing is optimized.