feat(mem): support multi-ported memories
Some designs may read/write multiple blocks in one clock cycle, such as
in a multi-channel DRAM system. To support this, this commit adds the
multi-ported feature for DifftestMem.

Note this changes the naming for blackbox and may break GSIM support.

We also move the DPI-C functions to src/test/vsrc because they should
not be declared multiple times, which may occur if multiple RWHelper's
are called.