feat(tracertl): generate 2-stage mmu page table
Add host page-table for every (g)paddr of the original page walk.
The new host pt doesn't change the guest page table. When the dut
disable 2-stage, only the guest page table is walked and the perf
result is the 1-stage result. So the hypervisor 2-stage mmu page
table is compatible with both 1-stage dut and 2-stage dut.

Signed-off-by: ZhangZifei <zhangzifei16@mails.ucas.ac.cn>