fix(DiffTestMem): import read/write dpic inside MemRWHelper
When MemRWHelper is instantiated multiple times with different
port numbers, different modules and files will be generated.
And import dpic statement outside module (global visible) will
cause conflicts.

Previously, we placed the import dpic statement in a separate file,
but it's easy to miss when compiling single file independently.

This change move the import dpic statement back into each file and
within the module to limit the scope of visibility.