feat(refill-check): add mask to refill check (#723)
After OpenXiangShan/XiangShan#3051, we have supported low-power 
read (read only 5 out of 8 8B-banks) in ICacheDataArray, and in V3, it was 
further modified to dynamically compute the mask based on the taken 
branch location.

Though, refill-check in difftest does not support masked comparison, 
instead it checks entire 512bit, so it will fail if we enable checks on 
cacheid >=3 (#712). As we're refactoring ICache recently, we need 
this check to ensure the refactored code is working properly.

Therefore, this PR tries to introduce masked refill-check, each mask 
bit controls comparison of 8B data.