fix(difftest): use ld as linker for BOLT PGO builds
When building with PGO, use `-fuse-ld=ld` to specify `ld` as the linker,
since mold from Ubuntu 22.04 does not support emitting relocations,
resulting in build failures if our LLVM is built with mold as the
default linker. While ld is still able to link all the objects in 1s on
my machine, that's fast enough for our use case.

Signed-off-by: Yangyu Chen <cyy@cyyself.name>