fix(verilator,gsim): use bfd linker for PGO+BOLT builds (#776)
When verilator is built to use GCC, current LDFLAGS use `-fuse-ld=ld`
will cause gcc error as `ld` is not recognized. Change it to use
`-fuse-ld=bfd` instead to compatible with GCC. Since bfd is the default
linker for most GNU Linux systems, this should not cause any change in
LLVM builds.

Some users may use verilator with GCC toolchain, so this change is
necessary to avoid build errors.

Signed-off-by: Yangyu Chen <cyy@cyyself.name>